Remove FFmbc
Compilation fails since MinGW upgrade to GCC 5.2.
If eventually it gets fixed again, just revert this commit, if not,
good riddance.

This used an ancient version of FFmpeg as base and because of that
it was always a gamble whether it would continue to work with the libs
it used always being updated.

In most of the cases, you should be using FFmpeg instead of FFmbc. If
there is some bug or problem you should report it to FFmpeg and they
will probably fix it.