[update.sh] Force install as explicit (fix #138)
In case some packages are installed but as dependencies this would
reinstall them but keep them as deps so it would always ask to reinstall.

Use this trick for removing packages too. If you don't want to remove them,
it will mark the packages as dependencies and never ask to remove them.