[ffmpeg] Don't compile with libvo-aacenc by default
ffmpeg's native aac encoder is considered better quality already
and soon will no longer need "-strict experimental" to be used.

https://trac.ffmpeg.org/wiki/Encode/HighQualityAudio#AudioencodersFFmpegcanuse
https://trac.ffmpeg.org/wiki/Encode/AAC#libvo_aacenc

We still keep vo-aacenc in case people want it. It was added
since FFmpeg's was the only free but lower quality alternative.
Same logic as not having libfaac as default, since libfdk-aac is also
non-free with GPL but higher quality.