update.sh: Refactor profile checking
Forgot to add LANG to the seds so the profile kept getting deleted again.
This way there's no need to change this code for every change to the first and last lines of the profiles.

Profile writing should also be done here so the code in .bat won't be run for now.
Later we'll probably move the profiles themselves to maybe helper.sh and this new function as well.