[rtmp] Use GnuTLS if license isn't nonfree
If --enable-gnutls was missing and license was (l)gpl openssl was being
used both for librtmp and rtmpdump.

[rtmp] Compile rtmpdump with GnuTLS if not nonfree

librtmp is LGPL so it can be compiled with OpenSSL but the binaries are
GPL so to be distributable they need to use GnuTLS (or PolarSSL). If
they were dynamically linked OpenSSL may be used with GPL binaries, but
since we always compile stuff statically, GnuTLS it is.

helper.sh: nit: separate librtmp checks

To resume:
if nonfree and --enable-openssl/--enable-gnutls are missing:
 rtmpdump is compiled with OpenSSL
 if --enable-librtmp is present, it will add --enable-openssl

You can still compile both rtmpdump and ffmpeg with gnutls if nonfree,
but you'll have to add --enable-gnutls or else the script will prefer
OpenSSL for both.