[FFmpeg] Don't force OpenSSL with librtmp nonfree builds
If OpenSSL was the only lib turning FFmpeg nonfree, this allows one
to keep it GPL.
Since librtmp is LGPL it can be used with OpenSSL, but FFmpeg won't use
OpenSSL directly so it shouldn't matter (IANAL).
We still add GnuTLS if not nonfree since GnuTLS will be built anyway
to be used in librtmp.