Add two more choices to FFmpeg license: (L)GPL version3
You will be prompted to choose again to prevent using the wrong license.

I decided to add this since I learned LGPLv3 is not compatible with GPLv2
so it does matter which version you choose.
mpv needs a new option if people want access to an extra feature (NNEDI3)
which was (maybe) based on LGPLv3. Since mpv is GPLv2.1, it means that
to have the feature it will need to be upgraded to GPLv3.