vapoursynth: Allow with 32-bit mpv too
Both libs are available whether Python 32-bit or 64-bit or both are
available. But the DLLs are only available if both Python versions are
installed.
Basically, if you want to use 32-bit Vapoursynth with 32-bit mpv you
need that version installed, but it shouldn't matter in the compilation.
Also, this assumes you install Vapoursynth to the default location.