ffmpeg: add patch to dynamically load cuda
Since the question of whether it should remain non-free or not is
still up in the air, leave the checks as if it were still non-free.

This patch allows the compilation and running of FFmpeg without having
the CUDA SDK installed when used for cuvid and cuda hwupload.

libnpp still requires CUDA SDK to compile and run, and is still
definitely non-free.