gnutls: switching to tmp-update-ax-code-coverage branch
They started discussion again and someone requested to bring code coverage up-to-the latest now instead of working around it. So the previous patch doesn't work anymore, the 3.x stable tree doesn't and master doesn't work either because we are running on the latest autoconf. So using this branch is currently the only way to build gnutls successfully. We can switch later back to stable or master as you wish.