SVT-HEVC: Files not installed (moved) correctly
This PR attempts to fix the SVT-HEVC DLL library not being moved to the `bin-video` folder. 

I removed the `Svt-Hevc` folder since the dp_cmakeinstall always placed the `SvtHevcEncApp.exe` in the `bin-video` folder. After the `do_cmakeinstall` I'm doing a `do_install` to copy the `libSvtHevcEnc.dll` DLL to the `bin-video` folder so that the suite is successful at running. Otherwise, it will fail and we have to press `Enter` to continue building.