VLC patches to allow compilation with the medialibrary master branch (#1629)
Added the following 2 patches:

* [vlc-corrected-patch](https://gist.github.com/moisespr123/2369978dc603ed1e67bdf7aba7304416): This is a rebase of the original patch by @1480c1 to make it apply to VLC's master branch.
* [medialib-patch](https://gist.github.com/moisespr123/70e16f70f5d016c3d0f0dacbf97526f0): This allows us to compile VLC using medialibrary master branch.

Note that VLC is still buggy. It will work, but will present an exception upon launching it. Press ignore and it will work. It has issues playing back videos (some doesn't open the video window. Other have distorted colors), but I'm sure this is not related to the above patches and is something else with the codecs.