Early Exit: Provide a way to cleanly exit the build at defined points
The build process can take several hours. In case of build errors,
one is usually tempted to fix the error and re-run the build, but
continuing on failed state with changes does not exactly guarantee
the same result (or even a successful result) as when re-builing
from scratch with those changes applied.

Having exit points like added by this commit allows to perform a
build incrementally and cleanly from point-to-point, when the
intermediate results are saved each time.

This is also very useful for setting up automated builds via GH
actions or DevOps where intermediate results are cached only when
no errors have occured.